requirements.txt
fastapi
uvicorn
pandas
pymongo
pydantic
python-dotenv
email-validator